+++ products/jbrowse-cli/src/commands/add-assembly.ts
@@ -1,379 +1,17 @@
-import fs from 'fs'
-import path from 'path'
 import { parseArgs } from 'util'
 
+import { debug, printHelp } from '../utils'
+
 import {
-  debug,
-  printHelp,
-  readInlineOrFileJson,
-  readJsonFile,
-  resolveFileLocation,
-  writeJsonFile,
-} from '../utils'
-
-import type { Assembly, Config, Sequence } from '../base'
-
-const { rename, copyFile, mkdir, symlink } = fs.promises
-
-function isValidJSON(string: string) {
-  try {
-    JSON.parse(string)
-    return true
-  } catch (error) {
-    return false
-  }
-}
-
-function guessSequenceType(sequence: string) {
-  if (
-    sequence.endsWith('.fa') ||
-    sequence.endsWith('.fna') ||
-    sequence.endsWith('.fasta')
-  ) {
-    return 'indexedFasta'
-  }
-  if (
-    sequence.endsWith('.fa.gz') ||
-    sequence.endsWith('.fna.gz') ||
-    sequence.endsWith('.fasta.gz')
-  ) {
-    return 'bgzipFasta'
-  }
-  if (sequence.endsWith('.2bit')) {
-    return 'twoBit'
-  }
-  if (sequence.endsWith('.chrom.sizes')) {
-    return 'chromSizes'
-  }
-  if (sequence.endsWith('.json')) {
-    return 'custom'
-  }
-  if (isValidJSON(sequence)) {
-    return 'custom'
-  }
-  throw new Error(
-    'Could not determine sequence type automatically, add --type to specify it',
-  )
-}
-
-function needLoadData(location: string) {
-  let locationUrl: URL | undefined
-  try {
-    locationUrl = new URL(location)
-  } catch (error) {
-    // ignore
-  }
-  if (locationUrl) {
-    return false
-  }
-  return true
-}
-
-async function loadData(
-  load: string,
-  filePaths: string[],
-  destination: string,
-) {
-  let locationUrl: URL | undefined
-  try {
-    locationUrl = new URL(filePaths[0]!)
-  } catch (error) {
-    // ignore
-  }
-
-  if (locationUrl) {
-    return false
-  }
-  switch (load) {
-    case 'copy': {
-      await Promise.all(
-        filePaths.map(async filePath => {
-          if (!filePath) {
-            return undefined
-          }
-          return copyFile(
-            filePath,
-            path.join(path.dirname(destination), path.basename(filePath)),
-          )
-        }),
-      )
-      return true
-    }
-    case 'symlink': {
-      await Promise.all(
-        filePaths.map(async filePath => {
-          if (!filePath) {
-            return undefined
-          }
-          return symlink(
-            path.resolve(filePath),
-            path.join(path.dirname(destination), path.basename(filePath)),
-          )
-        }),
-      )
-      return true
-    }
-    case 'move': {
-      await Promise.all(
-        filePaths.map(async filePath => {
-          if (!filePath) {
-            return undefined
-          }
-          return rename(
-            filePath,
-            path.join(path.dirname(destination), path.basename(filePath)),
-          )
-        }),
-      )
-      return true
-    }
-    case 'inPlace': {
-      return false
-    }
-  }
-  return false
-}
-
-async function getAssembly({
-  runFlags,
-  argsSequence,
-  target,
-}: {
-  runFlags: any
-  argsSequence: string
-  target: string
-}): Promise<Assembly> {
-  let sequence: Sequence
-
-  if (needLoadData(argsSequence) && !runFlags.load) {
-    throw new Error(
-      'Please specify the loading operation for this file with --load copy|symlink|move|inPlace',
-    )
-  } else if (!needLoadData(argsSequence) && runFlags.load) {
-    throw new Error(
-      'URL detected with --load flag. Please rerun the function without the --load flag',
-    )
-  }
-
-  let { name } = runFlags
-  let { type } = runFlags as {
-    type:
-      | 'indexedFasta'
-      | 'bgzipFasta'
-      | 'twoBit'
-      | 'chromSizes'
-      | 'custom'
-      | undefined
-  }
-  if (type) {
-    debug(`Type is: ${type}`)
-  } else {
-    type = guessSequenceType(argsSequence)
-    debug(`No type specified, guessing type: ${type}`)
-  }
-  if (name) {
-    debug(`Name is: ${name}`)
-  }
-  switch (type) {
-    case 'indexedFasta': {
-      const { skipCheck, force, load, faiLocation } = runFlags
-      let sequenceLocation = await resolveFileLocation(
-        argsSequence,
-        !(skipCheck || force),
-        load === 'inPlace',
-      )
-      debug(`FASTA location resolved to: ${sequenceLocation}`)
-      let indexLocation = await resolveFileLocation(
-        faiLocation || `${argsSequence}.fai`,
-        !(skipCheck || force),
-        load === 'inPlace',
-      )
-      debug(`FASTA index location resolved to: ${indexLocation}`)
-      if (!name) {
-        name = path.basename(
-          sequenceLocation,
-          sequenceLocation.endsWith('.fasta') ? '.fasta' : '.fa',
-        )
-        debug(`Guessing name: ${name}`)
-      }
-      const loaded = load
-        ? await loadData(load, [sequenceLocation, indexLocation], target)
-        : false
-      if (loaded) {
-        sequenceLocation = path.basename(sequenceLocation)
-        indexLocation = path.basename(indexLocation)
-      }
-      sequence = {
-        type: 'ReferenceSequenceTrack',
-        trackId: `${name}-ReferenceSequenceTrack`,
-        adapter: {
-          type: 'IndexedFastaAdapter',
-          fastaLocation: {
-            uri: sequenceLocation,
-            locationType: 'UriLocation',
-          },
-          faiLocation: { uri: indexLocation, locationType: 'UriLocation' },
-        },
-      }
-      break
-    }
-    case 'bgzipFasta': {
-      let sequenceLocation = await resolveFileLocation(
-        argsSequence,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(`compressed FASTA location resolved to: ${sequenceLocation}`)
-      let indexLocation = await resolveFileLocation(
-        runFlags.faiLocation || `${sequenceLocation}.fai`,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(`compressed FASTA index location resolved to: ${indexLocation}`)
-      let bgzipIndexLocation = await resolveFileLocation(
-        runFlags.gziLocation || `${sequenceLocation}.gzi`,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(`bgzip index location resolved to: ${bgzipIndexLocation}`)
-      if (!name) {
-        name = path.basename(
-          sequenceLocation,
-          sequenceLocation.endsWith('.fasta.gz') ? '.fasta.gz' : '.fa.gz',
-        )
-        debug(`Guessing name: ${name}`)
-      }
-      const loaded = runFlags.load
-        ? await loadData(
-            runFlags.load,
-            [sequenceLocation, indexLocation, bgzipIndexLocation],
-            target,
-          )
-        : false
-      if (loaded) {
-        sequenceLocation = path.basename(sequenceLocation)
-        indexLocation = path.basename(indexLocation)
-        bgzipIndexLocation = path.basename(bgzipIndexLocation)
-      }
-      sequence = {
-        type: 'ReferenceSequenceTrack',
-        trackId: `${name}-ReferenceSequenceTrack`,
-        adapter: {
-          type: 'BgzipFastaAdapter',
-          fastaLocation: {
-            uri: sequenceLocation,
-            locationType: 'UriLocation',
-          },
-          faiLocation: { uri: indexLocation, locationType: 'UriLocation' },
-          gziLocation: {
-            uri: bgzipIndexLocation,
-            locationType: 'UriLocation',
-          },
-        },
-      }
-      break
-    }
-    case 'twoBit': {
-      let sequenceLocation = await resolveFileLocation(
-        argsSequence,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(`2bit location resolved to: ${sequenceLocation}`)
-      if (!name) {
-        name = path.basename(sequenceLocation, '.2bit')
-        debug(`Guessing name: ${name}`)
-      }
-      const loaded = runFlags.load
-        ? await loadData(runFlags.load, [sequenceLocation], target)
-        : false
-      if (loaded) {
-        sequenceLocation = path.basename(sequenceLocation)
-      }
-      sequence = {
-        type: 'ReferenceSequenceTrack',
-        trackId: `${name}-ReferenceSequenceTrack`,
-        adapter: {
-          type: 'TwoBitAdapter',
-          twoBitLocation: {
-            uri: sequenceLocation,
-            locationType: 'UriLocation',
-          },
-        },
-      }
-      break
-    }
-    case 'chromSizes': {
-      let sequenceLocation = await resolveFileLocation(
-        argsSequence,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(`chrom.sizes location resolved to: ${sequenceLocation}`)
-      if (!name) {
-        name = path.basename(sequenceLocation, '.chrom.sizes')
-        debug(`Guessing name: ${name}`)
-      }
-      const loaded = runFlags.load
-        ? await loadData(runFlags.load, [sequenceLocation], target)
-        : false
-      if (loaded) {
-        sequenceLocation = path.basename(sequenceLocation)
-      }
-      sequence = {
-        type: 'ReferenceSequenceTrack',
-        trackId: `${name}-ReferenceSequenceTrack`,
-        adapter: {
-          type: 'ChromSizesAdapter',
-          chromSizesLocation: {
-            uri: sequenceLocation,
-            locationType: 'UriLocation',
-          },
-        },
-      }
-      break
-    }
-    case 'custom': {
-      const adapter = await readInlineOrFileJson<{ type: string }>(argsSequence)
-      debug(`Custom adapter: ${JSON.stringify(adapter)}`)
-      if (!name) {
-        if (isValidJSON(argsSequence)) {
-          throw new Error(
-            'Must provide --name when using custom inline JSON sequence',
-          )
-        } else {
-          name = path.basename(argsSequence, '.json')
-        }
-        debug(`Guessing name: ${name}`)
-      }
-      if (!('type' in adapter)) {
-        throw new Error(
-          `No "type" specified in sequence adapter "${JSON.stringify(
-            adapter,
-          )}"`,
-        )
-      }
-      sequence = {
-        type: 'ReferenceSequenceTrack',
-        trackId: `${name}-ReferenceSequenceTrack`,
-        adapter,
-      }
-      break
-    }
-  }
-
-  return { name, sequence }
-}
+  addAssemblyToConfig,
+  enhanceAssembly,
+  getAssembly,
+  loadOrCreateConfig,
+  resolveTargetPath,
+  saveConfigAndReport,
+} from './assembly-utils'
 
 export async function run(args?: string[]) {
-  // https://stackoverflow.com/a/35008327/2129219
-  const exists = (s: string) =>
-    new Promise(r => {
-      fs.access(s, fs.constants.F_OK, e => {
-        r(!e)
-      })
-    })
-
   const options = {
     type: {
       type: 'string',
@@ -509,128 +147,23 @@ export async function run(args?: string[]) {
   const argsSequence = positionals[0] || ''
   const output = runFlags.target || runFlags.out || '.'
 
-  if (!(await exists(output))) {
-    const dir = output.endsWith('.json') ? path.dirname(output) : output
-    await mkdir(dir, { recursive: true })
-  }
-  let isDir = false
-  try {
-    isDir = fs.statSync(output).isDirectory()
-  } catch (e) {}
-  const target = isDir ? path.join(output, 'config.json') : output
-
   debug(`Sequence location is: ${argsSequence}`)
-  const { name } = runFlags
 
-  const assembly = await getAssembly({ runFlags, argsSequence, target })
-  if (runFlags.alias?.length) {
-    debug(`Adding assembly aliases: ${runFlags.alias}`)
-    assembly.aliases = runFlags.alias
-  }
+  const target = await resolveTargetPath(output)
+  const baseAssembly = await getAssembly({ runFlags, argsSequence, target })
+  const assembly = await enhanceAssembly(baseAssembly, runFlags)
 
-  if (runFlags.refNameColors) {
-    assembly.refNameColors = runFlags.refNameColors
-      .split(',')
-      .map(color => color.trim())
-  }
+  const configContents = await loadOrCreateConfig(target)
+  const { config: updatedConfig, wasOverwritten } = await addAssemblyToConfig({
+    config: configContents,
+    assembly,
+    runFlags,
+  })
 
-  if (runFlags.refNameAliases) {
-    if (
-      runFlags.refNameAliasesType &&
-      runFlags.refNameAliasesType === 'custom'
-    ) {
-      const refNameAliasesConfig = await readInlineOrFileJson<{ type: string }>(
-        runFlags.refNameAliases,
-      )
-      if (!refNameAliasesConfig.type) {
-        throw new Error(
-          `No "type" specified in refNameAliases adapter "${JSON.stringify(
-            refNameAliasesConfig,
-          )}"`,
-        )
-      }
-      debug(
-        `Adding custom refNameAliases config: ${JSON.stringify(
-          refNameAliasesConfig,
-        )}`,
-      )
-      assembly.refNameAliases = {
-        adapter: refNameAliasesConfig,
-      }
-    } else {
-      const refNameAliasesLocation = await resolveFileLocation(
-        runFlags.refNameAliases,
-        !(runFlags.skipCheck || runFlags.force),
-        runFlags.load === 'inPlace',
-      )
-      debug(
-        `refName aliases file location resolved to: ${refNameAliasesLocation}`,
-      )
-      assembly.refNameAliases = {
-        adapter: {
-          type: 'RefNameAliasAdapter',
-          location: {
-            uri: refNameAliasesLocation,
-            locationType: 'UriLocation',
-          },
-        },
-      }
-    }
-  }
-
-  if (runFlags.displayName) {
-    assembly.displayName = runFlags.displayName
-  }
-
-  const defaultConfig: Config = {
-    assemblies: [],
-    configuration: {},
-    connections: [],
-    defaultSession: {
-      name: 'New Session',
-    },
-    tracks: [],
-  }
-
-  let configContents: Config
-
-  if (fs.existsSync(target)) {
-    debug(`Found existing config file ${target}`)
-    configContents = {
-      ...defaultConfig,
-      ...(await readJsonFile(target)),
-    }
-  } else {
-    debug(`Creating config file ${target}`)
-    configContents = { ...defaultConfig }
-  }
-
-  if (!configContents.assemblies) {
-    configContents.assemblies = []
-  }
-  const idx = configContents.assemblies.findIndex(
-    configAssembly => configAssembly.name === assembly.name,
-  )
-  if (idx !== -1) {
-    debug(`Found existing assembly ${name} in configuration`)
-    if (runFlags.overwrite || runFlags.force) {
-      debug(`Overwriting assembly ${name} in configuration`)
-      configContents.assemblies[idx] = assembly
-    } else {
-      throw new Error(
-        `Cannot add assembly with name ${assembly.name}, an assembly with that name already exists`,
-      )
-    }
-  } else {
-    configContents.assemblies.push(assembly)
-  }
-
-  debug(`Writing configuration to file ${target}`)
-  await writeJsonFile(target, configContents)
-
-  console.log(
-    `${idx !== -1 ? 'Overwrote' : 'Added'} assembly "${assembly.name}" ${
-      idx !== -1 ? 'in' : 'to'
-    } ${target}`,
-  )
+  await saveConfigAndReport({
+    config: updatedConfig,
+    target,
+    assembly,
+    wasOverwritten,
+  })
 }
