+++ products/jbrowse-cli/src/commands/add-track-utils/file-operations.ts
