+++ products/jbrowse-cli/src/commands/text-index-utils/aggregate.ts
