+++ products/jbrowse-cli/src/commands/text-index.ts
